+++ src/frontend/org/voltdb/sysprocs/SnapshotSave.java
@@ -87,7 +87,7 @@ public class SnapshotSave extends VoltSystemProcedure
                 // progress.
                 SNAP_LOG.error("@SnapshotSave is called while another snapshot is still in progress");
                 result.addRow(context.getHostId(), hostname, null, "FAILURE", "SNAPSHOT IN PROGRESS");
-                return new DependencyPair.TableDependencyPair((int) SysProcFragmentId.PF_createSnapshotTargets, result);
+                return new DependencyPair.TableDependencyPair(SysProcFragmentId.PF_createSnapshotTargets, result);
             }
 
             // Tell each site to quiesce - bring the Export and DR system to a steady state with
@@ -151,12 +151,12 @@ public class SnapshotSave extends VoltSystemProcedure
                     ((RealVoltDB)VoltDB.instance()).updateReplicaForJoin(context.getSiteId(), txnId);
                 }
             }
-            return new DependencyPair.TableDependencyPair((int) SysProcFragmentId.PF_createSnapshotTargets, result);
+            return new DependencyPair.TableDependencyPair(SysProcFragmentId.PF_createSnapshotTargets, result);
         }
         else if (fragmentId == SysProcFragmentId.PF_createSnapshotTargetsResults)
         {
-            VoltTable result = VoltTableUtil.unionTables(dependencies.get((int) SysProcFragmentId.PF_createSnapshotTargets));
-            return new DependencyPair.TableDependencyPair((int) SysProcFragmentId.PF_createSnapshotTargetsResults, result);
+            VoltTable result = VoltTableUtil.unionTables(dependencies.get(SysProcFragmentId.PF_createSnapshotTargets));
+            return new DependencyPair.TableDependencyPair(SysProcFragmentId.PF_createSnapshotTargetsResults, result);
         }
         assert (false);
         return null;
@@ -328,7 +328,7 @@ public class SnapshotSave extends VoltSystemProcedure
         long hashinatorVersion = (hashinatorData != null ? hashinatorData.m_version : 0);
         pfs[0] = new SynthesizedPlanFragment();
         pfs[0].fragmentId = SysProcFragmentId.PF_createSnapshotTargets;
-        pfs[0].outputDepId = (int) SysProcFragmentId.PF_createSnapshotTargets;
+        pfs[0].outputDepId = SysProcFragmentId.PF_createSnapshotTargets;
         pfs[0].multipartition = true;
         pfs[0].parameters = ParameterSet.fromArrayNoCopy(
                 filePath, fileNonce, txnId, perPartitionTxnIds, block, format.name(), data,
@@ -338,12 +338,12 @@ public class SnapshotSave extends VoltSystemProcedure
         // This fragment aggregates the results of creating those files
         pfs[1] = new SynthesizedPlanFragment();
         pfs[1].fragmentId = SysProcFragmentId.PF_createSnapshotTargetsResults;
-        pfs[1].outputDepId = (int) SysProcFragmentId.PF_createSnapshotTargetsResults;
+        pfs[1].outputDepId = SysProcFragmentId.PF_createSnapshotTargetsResults;
         pfs[1].multipartition = false;
         pfs[1].parameters = ParameterSet.emptyParameterSet();
 
         VoltTable[] results;
-        results = executeSysProcPlanFragments(pfs, (int) SysProcFragmentId.PF_createSnapshotTargetsResults);
+        results = executeSysProcPlanFragments(pfs, SysProcFragmentId.PF_createSnapshotTargetsResults);
         return results;
     }
 }
